fixed report dialog and added confirmation dialog for reset

+++ src/app/core/support/support/support.component.ts
@@ -1,7 +1,7 @@
 import { Component, Inject, OnInit } from "@angular/core";
 import { SessionService } from "../../session/session-service/session.service";
 import { DatabaseUser } from "../../session/session-service/local-user";
-import { WINDOW_TOKEN } from "../../../utils/di-tokens";
+import { LOCATION_TOKEN, WINDOW_TOKEN } from "../../../utils/di-tokens";
 import { SyncState } from "../../session/session-states/sync-state.enum";
 import { SwUpdate } from "@angular/service-worker";
 import { Database } from "../../database/database";
@@ -9,6 +9,7 @@ import * as Sentry from "@sentry/browser";
 import { Severity } from "@sentry/browser";
 import { RemoteSession } from "../../session/session-service/remote-session";
 import { RouteTarget } from "../../../app.routing";
+import { ConfirmationDialogService } from "../../confirmation-dialog/confirmation-dialog.service";
 
 @RouteTarget("Support")
 @Component({
@@ -27,9 +28,11 @@ export class SupportComponent implements OnInit {
 
   constructor(
     private sessionService: SessionService,
-    @Inject(WINDOW_TOKEN) private window: Window,
     private sw: SwUpdate,
-    private database: Database
+    private database: Database,
+    private confirmationDialog: ConfirmationDialogService,
+    @Inject(WINDOW_TOKEN) private window: Window,
+    @Inject(LOCATION_TOKEN) private location: Location
   ) {}
 
   ngOnInit(): void {
@@ -43,38 +46,42 @@ export class SupportComponent implements OnInit {
   private initCurrentSyncState() {
     switch (this.sessionService.syncState.value) {
       case SyncState.COMPLETED:
-        this.currentSyncState = "synced";
+        this.currentSyncState = $localize`:status|something is synced:synced`;
         return;
       case SyncState.STARTED:
-        this.currentSyncState = "in progress";
+        this.currentSyncState = $localize`:status|something is progress:in progress`;
         return;
       default:
-        this.currentSyncState = "unsynced";
+        this.currentSyncState = $localize`:status|something is unsynced:unsynced`;
     }
   }
 
   private initLastSync() {
     this.lastSync =
-      localStorage.getItem(SupportComponent.LAST_SYNC_KEY) || "never";
+      localStorage.getItem(SupportComponent.LAST_SYNC_KEY) ||
+      $localize`:timestamp|e.g. last sync - never:never`;
   }
 
   private initLastRemoteLogin() {
     this.lastRemoteLogin =
-      localStorage.getItem(RemoteSession.LAST_LOGIN_KEY) || "never";
+      localStorage.getItem(RemoteSession.LAST_LOGIN_KEY) ||
+      $localize`:timestamp|e.g. last sync - never:never`;
   }
 
   private initSwStatus() {
     if (this.sw.isEnabled) {
-      this.swStatus = "enabled";
+      this.swStatus = $localize`:status|something is enabled:enabled`;
     } else {
-      this.swStatus = "not enabled";
+      this.swStatus = $localize`:status|something is not enabled:not enabled`;
     }
   }
 
   sendReport() {
-    Sentry.addBreadcrumb({
-      type: "report information",
-      data: {
+    // This is sent even without submitting the crash report.
+    Sentry.captureMessage("report information", {
+      user: { name: this.currentUser.name },
+      level: Severity.Debug,
+      extra: {
         currentSyncState: this.currentSyncState,
         lastSync: this.lastSync,
         lastRemoteLogin: this.lastRemoteLogin,
@@ -85,15 +92,31 @@ export class SupportComponent implements OnInit {
     Sentry.showReportDialog({
       user: {
         name: this.currentUser.name,
+        email: "example@email.com",
       },
+      title: $localize`:Title user feedback dialog:Support request`,
+      subtitle: $localize`:Subtitle user feedback dialog:Please describe the problem you are facing.`,
+      subtitle2: "",
     });
   }
 
   async resetApplication() {
+    const choice = await this.confirmationDialog
+      .openDialog(
+        $localize`Reset Application`,
+        $localize`Are you sure you want to reset the application? This will delete all application data from your device and you will have to synchronize again.`
+      )
+      .afterClosed()
+      .toPromise();
+    if (!choice) {
+      return;
+    }
+
     await this.database.destroy();
     const registrations = await this.window.navigator.serviceWorker.getRegistrations();
     const unregisterPromises = registrations.map((reg) => reg.unregister());
     await Promise.all(unregisterPromises);
     localStorage.clear();
+    this.location.reload();
   }
 }
